WoT restructure A3: layout/ -> boxLayout/ + createLayoutDeps.ts -> gatherLayoutInput.ts
