+++ handsontable/src/3rdparty/walkontable/src/boxLayout/gatherLayoutInput.ts
