+++ handsontable/src/3rdparty/walkontable/src/boxLayout/layoutSnapshot.ts
