+++ handsontable/src/3rdparty/walkontable/src/boxLayout/resolveLayout.ts
